Renamed iptables-server.sh to iptables-standalone-server.sh.
Probably best to make a whole new script for a host which task is to act
as a gateway/router since forward rules and such needs to be added.
Feels bloated with forward rules if they're not going to be used.
If a good way is found to integrated it, it will be done.